+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/Tests.scala
@@ -41,7 +41,16 @@ object Tests {
       new PartyManagementServiceIT,
       new RaceConditionIT,
       new SemanticTests,
-      new TransactionServiceIT,
+      new TransactionServiceArgumentsIT,
+      new TransactionServiceAuthorizationIT,
+      new TransactionServiceCorrectnessIT,
+      new TransactionServiceExerciseIT,
+      new TransactionServiceOutputsIT,
+      new TransactionServiceQueryIT,
+      new TransactionServiceStakeholdersIT,
+      new TransactionServiceStreamsIT,
+      new TransactionServiceValidationIT,
+      new TransactionServiceVisibilityIT,
       new ValueLimitsIT,
       new WitnessesIT,
       new WronglyTypedContractIdIT,
